Fix problems with follow sets, epsilon, and the LR0 action table.
In the calculation of follow sets, the code may exit prematurely because it
may not recognize that it has made a change and should continue looping.

Epsilon shouldn't be added to the terminals because it is not a
terminal [1][2][3] and it also leads to epsilon being in the action table,
where it does not belong. Note that if it is no longer in the "terminals"
HashSet, there is a change needed in computeFirst to check for epsilon
explicitly.

There is another subtle problem at the bottom of computeFirst. The code
removes epsilon from the first set of a nullable intermediate non-terminal
and calls computeFirst recursively on the next nonterminal. Those are the
right things to do, but it does them in the wrong order. The recursive call
may legitimately return epsilon (if it is a last nonterminal with an
epsilon rule - see example above), but the code then removes it regardless.

In the closure of LR1State, there is a problem with computing the
next item for epsilon rules. When an epsilon item is created, it should be
"finished" immediately by placing the dot past the epsilon. This position
is where it should be so that REDUCE rules are created properly in
createActionTable(). Note that this is essentially what is suggested in [3]
except that they actually remove "epsilon" from the rule so that the
rule is more obviously completed.

Finally, there is a typo that results in an incorrect LR0 action table.

[1] https://cs.stackexchange.com/questions/85834/is-the-empty-string-a-terminal-symbol
[2] https://stackoverflow.com/questions/6501399/slr1-parser-and-epsilon-involved
[3] https://www.cs.uaf.edu/~cs331/notes/FirstFollow.pdf