Merge LR1 items that only differ by lookaheads.
This is done by updating the lookahead of the existing item.
Since the existing item is in the items HashSet and changing it will
change its hash code, we must remove it and add it back to the items
HashSet or else it will not be found later when checking for duplicate
states.